worker/requirements.txt
pytesseract==0.3.5
Pillow==7.2.0
